fix🐛: fix master certificate upload refresh issue
fixed a problem where after uploading master lisence, the image lists are not displaying correct
data. it turned out to be a re-rendering problem. I initially try to use the "imgUploadKey" as a
trigger to refresh, but forgot to actually bind this to any component. Here i ":key='imgUploadKey'"
to the image card div, and it successfully triggers re-render whenever imgUploadKey is changed,
which is the intented behaviour.